Redesigned the Tetrominos to make actions of the active Tetromino easier
A class per type of tetromino has now been added in contrast to the enums that were present earlier. The tetromino is now based on a fixed coordinate with each square drawn based on the location of said coordinate. All actions of the active tetromino is now located in the tetromino class. The active tetromino's squares will no longer be put into the tiles of the board when it's moving down, instead this will occur once the tetromino can't be moved down any further. Lastly functionality to increase the game's difficulty has been added